Add AlterEncryptRuleStatementUpdater
